Fix race in sensor tests.
There was a race for the notify channel because the channel was created
by the readLoop() goroutine, but read from the main test routine. The
reads didn't necessarily happen after `tkr` had been initialised, and
were therefore attempting to read from an uninitialised channel. This
lead to occasional test failures.

By initialising the channel in the test setup, this prevents the race.